[d3-chart] fixed pr issues
