+++ semcore/d3-chart/src/types/Plot.d.ts
@@ -103,9 +103,9 @@ export type PlotSummarizerConfig = {
    * **/
   additionalFields?: string[];
   /** Allows to format titles (e.g. axes) in alt text and data table */
-  titlesFormatter?: (key: string | number) => string | undefined;
+  titlesFormatter?: (key: string | number | null) => string | undefined;
   /** Allows to format values in alt text and data table */
-  valuesFormatter?: (value: unknown, key: string | number) => string | undefined;
+  valuesFormatter?: (value: unknown, key: string | number | null) => string | undefined;
 };
 
 /** @deprecated */
